% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbutils.R
\name{capwords}
\alias{capwords}
\title{Capitalise the first letter of each word in a string}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{A vector of one or more character strings}
}
\value{
A vector of one or more character strings with the first letter of each word capitalised
}
\description{
Capitalise the first letter of each word in a string
}
\examples{
# A single character string with two words
capwords('one two')
# Three character strings. The first has two words the last two
 have one word each
capwords(c('one two', 'three', 'four'))
}

