% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbutils.R
\name{mkdir}
\alias{mkdir}
\title{A simple function to create directories}
\usage{
mkdir(path)
}
\arguments{
\item{path}{A character string indicating the directory you want to create}
}
\description{
A simple function to create directories
}
\examples{
path <- 'C:/Users/rbirnie/Documents/example'
mkdir(path)

#This version creates all the subfolders down to the specified number of levels
path <- 'C:/Users/rbirnie/Documents/example/level2/level3/level4'
mkdir(path)
}

