% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbutils.R
\name{sas2csv}
\alias{sas2csv}
\title{Convert sas7bdat files to CSV}
\usage{
sas2csv(sas.dir, csv.dir)
}
\arguments{
\item{sas.dir}{Path to a directory full of SAS files in sas7bdat format}

\item{csv.dir}{Path to a directory where csv files should be saved. If this
directory does not exist it will be created automatically if possible}
}
\description{
Convert sas7bdat files to CSV
}
\details{
This function takes two arguments. A file path to a directory
  containing one or more SAS files and a file path to a directory where you
  would like the corresponding csv files to be saved. The conversin relies on
  the \code{read_sas} function from the haven package. You will need to
  install this if you do not have it already.
  \code{install.packages('haven')}
}
\seealso{
\code{\link[haven]{read_sas}}
}

