% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbutils.R
\name{saveXLSX}
\alias{saveXLSX}
\title{A simple wrapper around \code{write.xlsx} from the \code{xlsx} package}
\usage{
saveXLSX(x, file, sheetName = "Sheet1", col.names = TRUE,
  row.names = TRUE, append = FALSE, showNA = FALSE, overwrite = TRUE)
}
\arguments{
\item{x}{A \code{data.frame} to be saved as a single sheet in an excel file}

\item{file}{the path to the output file. This should include the '.xlsx' file extension}

\item{sheetName}{a character string with the sheet name.}

\item{col.names}{a logical value indicating if the column names of \code{x} are to be written along with \code{x} to the file.}

\item{row.names}{a logical value indicating whether the row names of \code{x} are to be written along with \code{x} to the file.}

\item{append}{a logical value indicating if \code{x} should be appended to an existing file. If \code{TRUE} the file is read from disk.}

\item{showNA}{a logical value. If set to FALSE, NA values will be left as empty cells}

\item{overwrite}{a logical value indicating if \code{x} should overwrite the contents of an existing sheet in the workbook}
}
\description{
A simple wrapper around \code{write.xlsx} from the \code{xlsx} package
}
\details{
Add a more verbose explanation here
}
\examples{
Add some examples. Use some existing R data sets
}
\seealso{
\code{\link[xlsx]{write.xlsx}}
}

