% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbutils.R
\name{saveXLSX}
\alias{saveXLSX}
\title{A simple wrapper around \code{write.xlsx} from the \code{xlsx} package}
\usage{
saveXLSX(x, file, sheetName = "Sheet1", col.names = TRUE,
  row.names = FALSE, append = FALSE, showNA = FALSE, overwrite = TRUE)
}
\arguments{
\item{x}{A \code{data.frame} to be saved as a single sheet in an excel file}

\item{file}{the path to the output file. This should include the '.xlsx' file
extension}

\item{sheetName}{a character string with the sheet name.}

\item{col.names}{a logical value indicating if the column names of \code{x}
are to be written along with \code{x} to the file.}

\item{row.names}{a logical value indicating whether the row names of \code{x}
are to be written along with \code{x} to the file.}

\item{append}{a logical value indicating if \code{x} should be appended to an
existing file. If \code{TRUE} the file is read from disk.}

\item{showNA}{a logical value. If set to FALSE, NA values will be left as
empty cells}

\item{overwrite}{a logical value indicating if \code{x} should overwrite the
  contents of an existing sheet in the workbook}
}
\description{
A simple wrapper around \code{write.xlsx} from the \code{xlsx} package
}
\details{
This function is a basic wrapper around \code{write.xlsx} from the
  \code{xlsx} package designed to modify some of the default behaviour. The
  main difference is the addition of the \code{overwrite} argument. This is
  intended to allow the user to overwrite sheets that already exist in the
  workbook. If the Excel file specified by \code{file} already exists and
  includes a worksheet with same name as \code{sheetName} then this sheet
  will be deleted and replaced if \code{overwrite=TRUE}. If
  \code{overwrite=FALSE} then the sheet is appended to the end of the file.

  The default for \code{append} in the underlying \code{write.xlsx} function
  is \code{FALSE}. If the file you want to write to already exists you should
  set this to \code{append=TRUE}. If the file exists and \code{append=FALSE}
  the existing file will be deleted and replaced. In most cases
  \code{append=TRUE} is probably what you want.
}
\examples{
#basic usage for a new file
#creates a temporary file in the current directory
data(iris)
f = paste0('./temp','.xlsx')
saveXLSX(x = iris, file = f, sheetName = 'Test')

#using the same file as above
#use a different data set but keep the same name
#if the file already exists set overwrite = TRUE and append = TRUE
data(CO2)
saveXLSX(x = CO2, file = f, sheetName = 'Test', append = TRUE,
overwrite=TRUE)

#if overwrite = FALSE then the new sheet is appended at
#the end of the file
#use another different data set.
data(Indometh)
saveXLSX(x = Indometh, file = f, sheetName = 'Test', append = TRUE,
overwrite=FALSE)
}
\seealso{
\code{\link[xlsx]{write.xlsx}}
}

